WIP refactoring
- product -> composite
- sum -> variant
- named and unnamed fields builder
