+++ src/impls.rs
@@ -95,7 +95,7 @@ where
 	T: Metadata + 'static,
 {
 	fn type_info() -> Type {
-		Type::r#struct("Result", Namespace::prelude())
+		Type::composite("Result", Namespace::prelude())
 			.type_params(tuple_meta_type![T])
 			.with_named_fields()
 			.named_field_of::<[T]>("elems")
