+++ src/type/composite.rs
@@ -17,179 +17,99 @@
 use crate::tm_std::*;
 
 use crate::{
-	form::{CompactForm, Form, MetaForm},
-	utils::is_rust_identifier,
-	IntoCompact, MetaType, Registry,
+	fields::{Field, NoFields, NamedFields},
+	form::{CompactForm, Form, MetaForm}, IntoCompact, Field, Path, Namespace, Registry
 };
+use derive_more::From;
 use serde::Serialize;
+use crate::fields::UnnamedFields;
 
-/// Represents a type composed from other types e.g. structs, enums
-#[derive(PartialEq, Eq, PartialOrd, Ord, Clone, Serialize, Debug)]
-#[serde(bound = "T: Serialize, F::TypeId: Serialize")]
-pub struct TypeComposite<T, F: Form = MetaForm> {
-	/// The name of the composite type.
-	name: F::String,
-	/// The namespace in which the composite type has been defined.
-	namespace: Namespace<F>,
-	/// The generic type parameters of the composite type in use.
-	#[serde(rename = "params")]
-	type_params: Vec<F::TypeId>,
-	/// The definition of the composite type
-	def: T,
+/// A composite type, consisting of either named (struct) or unnamed (tuple struct) fields
+///
+/// # Examples
+///
+/// ## A Rust struct with named fields.
+///
+/// ```
+/// struct Person {
+///     name: String,
+///     age_in_years: u8,
+///     friends: Vec<Person>,
+/// }
+/// ```
+///
+/// ## A tuple struct with unnamed fields.
+///
+/// ```
+/// struct Color(u8, u8, u8);
+/// ```
+///
+/// ## A so-called unit struct
+///
+/// ```
+/// struct JustAMarker;
+/// ```
+#[derive(PartialEq, Eq, PartialOrd, Ord, Clone, Debug, Serialize, From)]
+#[serde(bound = "F::TypeId: Serialize")]
+#[serde(rename_all = "lowercase")]
+pub struct TypeComposite<F: Form = MetaForm> {
+	path: Path<F>,
+	fields: Vec<T>,
 }
 
-impl<T: IntoCompact> IntoCompact for TypeComposite<T> {
-	type Output = TypeComposite<T::Output, CompactForm>;
+impl<T> IntoCompact for TypeComposite<T> {
+	type Output = TypeComposite<CompactForm>;
 
 	fn into_compact(self, registry: &mut Registry) -> Self::Output {
 		TypeComposite {
-			name: registry.register_string(self.name),
-			namespace: self.namespace.into_compact(registry),
-			type_params: self
-				.type_params
-				.into_iter()
-				.map(|param| registry.register_type(&param))
-				.collect::<Vec<_>>(),
-			def: self.def.into_compact(registry),
+			path: self.path.into_compact(),
+			fields: registry.map_into_compact(self.fields),
 		}
 	}
 }
 
 impl<T> TypeComposite<T> {
-	/// Creates a new type identifier to refer to a custom type definition.
-	pub fn new<P>(name: &'static str, namespace: Namespace, type_params: P, def: T) -> Self
-	where
-		P: IntoIterator<Item = MetaType>,
+	/// Creates a new struct definition with named fields.
+	pub fn new(name: &'static str, namespace: Namespace) -> TypeCompositeBuilder
 	{
+		TypeCompositeBuilder::new(
+			Self {
+				path: Path::new(name, namespace, Vec::new()),
+				fields: Vec::new(),
+			}
+		)
+	}
+
+	/// Creates the unit tuple-struct that has no fields.
+	pub fn unit(path: Path) -> Self {
+		Self::new(path).done()
+	}
+}
+
+pub struct TypeCompositeBuilder<F = NoFields> {
+	ty: TypeComposite,
+	fields_marker: PhantomData<fn() -> F>,
+}
+
+impl TypeCompositeBuilder {
+	pub fn new<F>(ty: TypeComposite) -> TypeCompositeBuilder<F> {
 		Self {
-			name,
-			namespace,
-			type_params: type_params.into_iter().collect(),
-			def,
+			ty,
+			fields_marker: Default::default()
 		}
 	}
-}
 
-/// Represents the namespace of a type definition.
-///
-/// This consists of several segments that each have to be a valid Rust identifier.
-/// The first segment represents the crate name in which the type has been defined.
-///
-/// Rust prelude type may have an empty namespace definition.
-#[derive(PartialEq, Eq, PartialOrd, Ord, Clone, Serialize, Debug)]
-#[serde(transparent)]
-pub struct Namespace<F: Form = MetaForm> {
-	/// The segments of the namespace.
-	segments: Vec<F::String>,
-}
-
-/// An error that may be encountered upon constructing namespaces.
-#[derive(PartialEq, Eq, Debug)]
-pub enum NamespaceError {
-	/// If the module path does not at least have one segment.
-	MissingSegments,
-	/// If a segment within a module path is not a proper Rust identifier.
-	InvalidIdentifier {
-		/// The index of the errorneous segment.
-		segment: usize,
-	},
-}
-
-impl IntoCompact for Namespace {
-	type Output = Namespace<CompactForm>;
-
-	/// Compacts this namespace using the given registry.
-	fn into_compact(self, registry: &mut Registry) -> Self::Output {
-		Namespace {
-			segments: self
-				.segments
-				.into_iter()
-				.map(|seg| registry.register_string(seg))
-				.collect::<Vec<_>>(),
-		}
-	}
-}
-
-impl Namespace {
-	/// Creates a new namespace from the given segments.
-	pub fn new<S>(segments: S) -> Result<Self, NamespaceError>
-	where
-		S: IntoIterator<Item = <MetaForm as Form>::String>,
-	{
-		let segments = segments.into_iter().collect::<Vec<_>>();
-		if segments.is_empty() {
-			return Err(NamespaceError::MissingSegments);
-		}
-		if let Some(err_at) = segments.iter().position(|seg| !is_rust_identifier(seg)) {
-			return Err(NamespaceError::InvalidIdentifier { segment: err_at });
-		}
-		Ok(Self { segments })
+	pub fn named_fields(self) -> TypeCompositeBuilder<NamedFields> {
+		Self::new(self.ty)
 	}
 
-	/// Creates a new namespace from the given module path.
-	///
-	/// # Note
-	///
-	/// Module path is generally obtained from the `module_path!` Rust macro.
-	pub fn from_module_path(module_path: <MetaForm as Form>::String) -> Result<Self, NamespaceError> {
-		Self::new(module_path.split("::"))
+	pub fn unnamed_fields(self) -> TypeCompositeBuilder<UnnamedFields> {
+		Self::new(self.ty)
 	}
 
-	/// Creates the prelude namespace.
-	pub fn prelude() -> Self {
-		Self { segments: vec![] }
-	}
-}
+	// todo: [AJ] add type params (only allow on types with fields?)
 
-#[cfg(test)]
-mod tests {
-	use super::*;
-
-	#[test]
-	fn namespace_ok() {
-		assert_eq!(
-			Namespace::new(vec!["hello"]),
-			Ok(Namespace {
-				segments: vec!["hello"]
-			})
-		);
-		assert_eq!(
-			Namespace::new(vec!["Hello", "World"]),
-			Ok(Namespace {
-				segments: vec!["Hello", "World"]
-			})
-		);
-		assert_eq!(Namespace::new(vec!["_"]), Ok(Namespace { segments: vec!["_"] }));
-	}
-
-	#[test]
-	fn namespace_err() {
-		assert_eq!(Namespace::new(vec![]), Err(NamespaceError::MissingSegments));
-		assert_eq!(
-			Namespace::new(vec![""]),
-			Err(NamespaceError::InvalidIdentifier { segment: 0 })
-		);
-		assert_eq!(
-			Namespace::new(vec!["1"]),
-			Err(NamespaceError::InvalidIdentifier { segment: 0 })
-		);
-		assert_eq!(
-			Namespace::new(vec!["Hello", ", World!"]),
-			Err(NamespaceError::InvalidIdentifier { segment: 1 })
-		);
-	}
-
-	#[test]
-	fn namespace_from_module_path() {
-		assert_eq!(
-			Namespace::from_module_path("hello::world"),
-			Ok(Namespace {
-				segments: vec!["hello", "world"]
-			})
-		);
-		assert_eq!(
-			Namespace::from_module_path("::world"),
-			Err(NamespaceError::InvalidIdentifier { segment: 0 })
-		);
+	pub fn done(self) -> TypeComposite {
+		self.ty
 	}
 }
